Fix potential bug when calling readline() without awaiting previous value
It was buggy if you tried to use it in any way other than from the
iterator, since if it was called while `more_` was still running it'd see
that the buffer was still empty and call `more_` again, causing lost items
and general confusion.

Inline the original body into the iterator (which is the only thing that
should be using it like that anyway) and replace it with an implementation
that advances the iterator and reformats the result.